feat(infisical-pg): updated pg migrator to new table name changes

+++ backend-pg/src/db/migrations/20231228075011_identity-access-token.ts
@@ -8,10 +8,10 @@ export async function up(knex: Knex): Promise<void> {
     await knex.schema.createTable(TableName.IdentityAccessToken, (t) => {
       t.uuid("id", { primaryKey: true }).defaultTo(knex.fn.uuid());
       t.string("authType").notNullable();
-      t.integer("accessTokenTTL").defaultTo(2592000).notNullable(); // 30 days second
-      t.integer("accessTokenMaxTTL").defaultTo(2592000).notNullable();
-      t.integer("accessTokenNumUses").defaultTo(0).notNullable();
-      t.integer("accessTokenNumUsesLimit").defaultTo(0).notNullable();
+      t.bigInteger("accessTokenTTL").defaultTo(2592000).notNullable(); // 30 days second
+      t.bigInteger("accessTokenMaxTTL").defaultTo(2592000).notNullable();
+      t.bigInteger("accessTokenNumUses").defaultTo(0).notNullable();
+      t.bigInteger("accessTokenNumUsesLimit").defaultTo(0).notNullable();
       t.datetime("accessTokenLastUsedAt");
       t.datetime("accessTokenLastRenewedAt");
       t.boolean("isAccessTokenRevoked").defaultTo(false).notNullable();
